pnorm formulation in cusum

+++ src/RcppExports.cpp
@@ -350,26 +350,26 @@ BEGIN_RCPP
 END_RCPP
 }
 // compute_norm
-NumericVector compute_norm(NumericMatrix x, int norm, bool display_progress);
+NumericVector compute_norm(NumericMatrix x, double norm, bool display_progress);
 RcppExport SEXP _swatanomaly_compute_norm(SEXP xSEXP, SEXP normSEXP, SEXP display_progressSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type norm(normSEXP);
+    Rcpp::traits::input_parameter< double >::type norm(normSEXP);
     Rcpp::traits::input_parameter< bool >::type display_progress(display_progressSEXP);
     rcpp_result_gen = Rcpp::wrap(compute_norm(x, norm, display_progress));
     return rcpp_result_gen;
 END_RCPP
 }
 // detect_norm
-LogicalVector detect_norm(NumericMatrix x, int norm, double threshold, bool display_progress);
+LogicalVector detect_norm(NumericMatrix x, double norm, double threshold, bool display_progress);
 RcppExport SEXP _swatanomaly_detect_norm(SEXP xSEXP, SEXP normSEXP, SEXP thresholdSEXP, SEXP display_progressSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type norm(normSEXP);
+    Rcpp::traits::input_parameter< double >::type norm(normSEXP);
     Rcpp::traits::input_parameter< double >::type threshold(thresholdSEXP);
     Rcpp::traits::input_parameter< bool >::type display_progress(display_progressSEXP);
     rcpp_result_gen = Rcpp::wrap(detect_norm(x, norm, threshold, display_progress));
@@ -377,7 +377,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // compute_cusum
-NumericVector compute_cusum(NumericMatrix x, int win, int jump, int norm, bool display_progress);
+NumericVector compute_cusum(NumericMatrix x, int win, int jump, double norm, bool display_progress);
 RcppExport SEXP _swatanomaly_compute_cusum(SEXP xSEXP, SEXP winSEXP, SEXP jumpSEXP, SEXP normSEXP, SEXP display_progressSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -385,14 +385,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
     Rcpp::traits::input_parameter< int >::type win(winSEXP);
     Rcpp::traits::input_parameter< int >::type jump(jumpSEXP);
-    Rcpp::traits::input_parameter< int >::type norm(normSEXP);
+    Rcpp::traits::input_parameter< double >::type norm(normSEXP);
     Rcpp::traits::input_parameter< bool >::type display_progress(display_progressSEXP);
     rcpp_result_gen = Rcpp::wrap(compute_cusum(x, win, jump, norm, display_progress));
     return rcpp_result_gen;
 END_RCPP
 }
 // detect_cusum
-LogicalVector detect_cusum(NumericMatrix x, int win, int jump, int norm, double threshold, bool display_progress);
+LogicalVector detect_cusum(NumericMatrix x, int win, int jump, double norm, double threshold, bool display_progress);
 RcppExport SEXP _swatanomaly_detect_cusum(SEXP xSEXP, SEXP winSEXP, SEXP jumpSEXP, SEXP normSEXP, SEXP thresholdSEXP, SEXP display_progressSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -400,7 +400,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
     Rcpp::traits::input_parameter< int >::type win(winSEXP);
     Rcpp::traits::input_parameter< int >::type jump(jumpSEXP);
-    Rcpp::traits::input_parameter< int >::type norm(normSEXP);
+    Rcpp::traits::input_parameter< double >::type norm(normSEXP);
     Rcpp::traits::input_parameter< double >::type threshold(thresholdSEXP);
     Rcpp::traits::input_parameter< bool >::type display_progress(display_progressSEXP);
     rcpp_result_gen = Rcpp::wrap(detect_cusum(x, win, jump, norm, threshold, display_progress));
